this is a much more solid foundation for listing
and viewing blogs.

So we don't get toggling blog expansion. Psh. I'd probably be opening up a blog in a
new tab anyway so I can focus on it while going back to the main blog list
to browse more articles.

Overall, a worthy tradeoff.

And there's no extra rendering.